conecte_me/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
requests==2.31.0
python-decouple==3.8  # Ajouté pour une meilleure gestion des variables d'environnement
django-oauth-toolkit==2.0.0


